Hook up the front-end to the back-end in a very simple way.
