+++ mobile/package.json
@@ -19,8 +19,14 @@
     "preset": "jest-expo"
   },
   "dependencies": {
+    "apollo-cache-inmemory": "^1.1.5",
+    "apollo-client": "^2.2.0",
+    "apollo-link-http": "^1.3.2",
     "expo": "^23.0.4",
+    "graphql": "^0.12.3",
+    "graphql-tag": "^2.6.1",
     "react": "16.0.0",
+    "react-apollo": "^2.0.4",
     "react-native": "0.50.3",
     "react-native-elements": "^0.18.5"
   }
